Fixed Errors, and Reorganize work contribution in the userprofile module
