+++ app/src/main/java/com/android/tripbook/MainActivity.kt
@@ -13,14 +13,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.android.tripbook.ui.theme.TripBookTheme
 import com.android.tripbook.ui.UserProfileScreen // Import the UserProfileScreen function
-//import com.tripbook.userprofilendedilan.UserProfileNdeDilanEntryPoint
+import com.tripbook.userprofilendedilan.UserProfileNdeDilanEntryPoint
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            //UserProfileNdeDilanEntryPoint( )
+            UserProfileNdeDilanEntryPoint( )
             TripBookTheme {
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
                     // Calling the UserProfileScreen composable
