+++ userprofile/src/main/java/com/tripbook/userprofileTalaDaryll/database/AppDatabase.kt
@@ -1,9 +1,10 @@
-package com.android.tripbook
+package com.tripbook.userprofileTalaDaryll.database
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.tripbook.userprofileTalaDaryll.models.Souvenir
 
 @Database(entities = [Souvenir::class], version = 1, exportSchema = false)
 abstract class AppDatabase : RoomDatabase() {
