+++ userprofile/src/main/java/com/tripbook/userprofileTalaDaryll/database/SouvenirRepository.kt
@@ -1,6 +1,7 @@
-package com.android.tripbook
+package com.tripbook.userprofileTalaDaryll.database
 
 import kotlinx.coroutines.flow.Flow
+import com.tripbook.userprofileTalaDaryll.models.Souvenir
 
 class SouvenirRepository(private val souvenirDao: SouvenirDao) {
     val allSouvenirs: Flow<List<Souvenir>> = souvenirDao.getAllSouvenirs()
