+++ userprofile/src/main/java/com/tripbook/userprofileTalaDaryll/database/TripBookApplication.kt
@@ -1,0 +1,12 @@
+package com.tripbook.userprofileTalaDaryll.database
+
+import android.app.Application
+import com.google.firebase.FirebaseApp
+
+class TripBookApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        FirebaseApp.initializeApp(this)
+    }
+}
+
