+++ userprofile/src/main/res/xml/file_paths.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<paths>
+<paths xmlns:android="http://schemas.android.com/apk/res/android">
     <external-files-path
         name="profile_pictures"
         path="profile_pictures" />
+    <files-path name="files" path="." />
+    <external-path name="external_files" path="." />
 </paths>
